chore: commit current workspace

+++ backend/agent-persistence-postgres/src/main/java/org/congcong/algomentor/agent/persistence/postgres/mapper/AgentContextSnapshotMapper.java
@@ -6,5 +6,5 @@ import org.congcong.algomentor.agent.persistence.postgres.mapper.model.ContextSn
 @Mapper
 public interface AgentContextSnapshotMapper {
 
-  int insertSnapshot(ContextSnapshotRow row);
+  long insertSnapshot(ContextSnapshotRow row);
 }
